fix: Change the type of timeout from u128 to u64

+++ src/record/settings.rs
@@ -4,7 +4,7 @@ use clap::ArgMatches;
 #[derive(Debug)]
 pub struct RecordSettings {
 	pub fps: u32,
-	pub timeout: u128,
+	pub timeout: u64,
 	pub interval: u64,
 	pub countdown: u64,
 	pub color: u64,
@@ -13,7 +13,7 @@ pub struct RecordSettings {
 impl RecordSettings {
 	pub fn new(
 		fps: u32,
-		timeout: u128,
+		timeout: u64,
 		interval: u64,
 		countdown: u64,
 		color: u64,
@@ -33,7 +33,7 @@ impl RecordSettings {
 				let settings_parser = SettingsParser::new(matches.clone());
 				Self::new(
 					settings_parser.get_arg::<u32>("fps", 10),
-					settings_parser.get_arg::<u128>("timeout", 30),
+					settings_parser.get_arg::<u64>("timeout", 30),
 					settings_parser.get_arg::<u64>("interval", 10),
 					settings_parser.get_arg::<u64>("countdown", 3),
 					color,
